[GCOM-1108] rename package
