+++ packages/hygraph-dynamic-rows-ui/lib/fetchGraphQLInterface.ts
