+++ packages/hygraph-dynamic-rows-ui/lib/objectifyGraphQLInterface.ts
